Fixes issue with missing licence headers
